refactor(uploads): extract arcgis commit orchestration for worker reuse
